Moved metadata fields and templates
