+++ src/main/resources/i18n/messages.properties
@@ -580,7 +580,6 @@ project.associated.table.associated=Associated
 project.nav.analysis=Analyses
 project.nav.samples=Samples
 project.nav.linelist=Line List
-project.nav.metadata=Sample Metadata
 project.nav.members=Members
 project.nav.associated=Associated Projects
 project.nav.referenceFiles=Reference Files
@@ -954,6 +953,7 @@ project.settings.page.processing=Processing
 project.settings.page.remote=Remote
 project.settings.page.members=Members
 project.settings.page.groups=Groups
+project.settings.page.metadata=Metadata
 project.settings.page.referenceFiles=Reference Files
 project.settings.page.associated=Associated Projects
 project.settings.metadata-templates=Metadata
