+++ src/main/webapp/entries.js
@@ -62,7 +62,6 @@ module.exports = {
   search: "./resources/js/pages/search/search.js",
   "run-files": "./resources/js/pages/sequence-files/run-files.js",
   "user-details": "./resources/js/pages/user-details.js",
-  "project-metadata": "./resources/js/pages/projects/metadata",
   admin: "./resources/js/pages/admin/index.js",
   "sequence-files": "./resources/js/pages/sequence-files/",
 };
