+++ src/main/webapp/resources/js/components/project/ProjectNav.jsx
@@ -1,6 +1,6 @@
+import { Layout, Menu, PageHeader } from "antd";
 import React, { useState } from "react";
 import { render } from "react-dom";
-import { Layout, Menu, PageHeader } from "antd";
 import { setBaseUrl } from "../../utilities/url-utilities";
 import { IconFolder } from "../icons/Icons";
 import { RemoteProjectStatus } from "./RemoteProjectStatus";
@@ -33,11 +33,6 @@ export function ProjectNav() {
           </Item>
           <Item key="linelist">
             <a href={`${BASE_URL}linelist`}>{i18n("project.nav.linelist")}</a>
-          </Item>
-          <Item key="metadata">
-            <a href={`${BASE_URL}metadata/fields`}>
-              {i18n("project.nav.metadata")}
-            </a>
           </Item>
           <Item key="analyses">
             <a href={`${BASE_URL}analyses`}>{i18n("project.nav.analysis")}</a>
