+++ src/main/webapp/resources/js/pages/projects/settings/components/SettingsNav.jsx
@@ -23,6 +23,19 @@ export default function SettingsNav({ path }) {
       <Menu.Item key="groups">
         <Link to="groups">{i18n("project.settings.page.groups")}</Link>
       </Menu.Item>
+      <Menu.SubMenu
+        key="metadata"
+        title={i18n("project.settings.page.metadata")}
+      >
+        <Menu.Item key="metadata-fields">
+          <Link to="metadata-fields">{i18n("MetadataFields.title")}</Link>
+        </Menu.Item>
+        <Menu.Item key="metadata-templates">
+          <Link to="metadata-templates">
+            {i18n("ProjectMetadataTemplates.title")}
+          </Link>
+        </Menu.Item>
+      </Menu.SubMenu>
     </Menu>
   );
 }
