+++ src/main/webapp/resources/js/pages/projects/settings/components/metadata/MetadataFields.jsx
