+++ src/main/webapp/resources/js/pages/projects/settings/components/metadata/MetadataFieldsList.jsx
@@ -1,9 +1,9 @@
+import { Button, Empty, Space, Table, Typography } from "antd";
 import React from "react";
-import { Button, Empty, Space, Table } from "antd";
+import { useDispatch, useSelector } from "react-redux";
+import { fetchFieldsForProject } from "../../../redux/fieldsSlice";
+import { fetchTemplatesForProject } from "../../../redux/templatesSlice";
 import { MetadataTemplateCreate } from "./MetadataTemplateCreate";
-import { useSelector } from "react-redux";
-import { SPACE_MD } from "../../../styles/spacing";
-import { IconFolder } from "../../../components/icons/Icons";
 
 /**
  * Component for showing metadata fields associated with a project.
@@ -12,10 +12,16 @@ import { IconFolder } from "../../../components/icons/Icons";
  */
 export function MetadataFieldsList({ projectId }) {
   const { canManage } = useSelector((state) => state.project);
+  const dispatch = useDispatch();
   const { fields, loading } = useSelector((state) => state.fields);
 
   const [selected, setSelected] = React.useState([]);
   const [selectedFields, setSelectedFields] = React.useState([]);
+
+  React.useEffect(() => {
+    dispatch(fetchFieldsForProject(projectId));
+    dispatch(fetchTemplatesForProject(projectId));
+  }, [dispatch, projectId]);
 
   React.useEffect(() => {
     /*
@@ -44,6 +50,9 @@ export function MetadataFieldsList({ projectId }) {
 
   return (
     <Space direction="vertical" style={{ display: "block" }}>
+      <Typography.Title level={2}>
+        {i18n("MetadataFields.title")}
+      </Typography.Title>
       {canManage && (
         <Space>
           <MetadataTemplateCreate fields={selectedFields} projectId={projectId}>
