+++ src/main/webapp/resources/js/pages/projects/settings/components/metadata/MetadataTemplateMember.jsx
@@ -9,7 +9,7 @@ import {
 } from "antd";
 import React from "react";
 import { useSelector } from "react-redux";
-import { addKeysToList } from "../../../utilities/http-utilities";
+import { addKeysToList } from "../../../../../utilities/http-utilities";
 
 const { Paragraph, Text } = Typography;
