+++ src/main/webapp/resources/js/pages/projects/settings/store.js
@@ -1,7 +1,9 @@
 import { configureStore } from "@reduxjs/toolkit";
 import coverageReducer from "../redux/coverageSlice";
+import fieldsReducer from "../redux/fieldsSlice";
 import pipelineReducer from "../redux/pipelinesSlice";
 import projectReducer from "../redux/projectSlice";
+import templateReducer from "../redux/templatesSlice";
 import userReducer from "../redux/userSlice";
 
 /*
@@ -14,5 +16,7 @@ export default configureStore({
     coverage: coverageReducer,
     pipelines: pipelineReducer,
     user: userReducer,
+    fields: fieldsReducer,
+    templates: templateReducer,
   },
 });
